Forecasts based on DRM.

Considering a new dataset (across the same patches), computes forecasts based on the DRM passed as drm.

Usage

predict_drm(
  drm,
  drm_data,
  ntime_for,
  x_tt,
  x_rt,
  x_mt = matrix(1, ncol = 1),
  x_mpast = matrix(1, ncol = 1),
  f_test,
  seed = 1,
  cores = 1
)

Arguments

drm

A CmdStanFit object containing samples from the posterior distribution.

drm_data

a list used as input for model fitting. Typically, the output from the make_data function.

ntime_for

an integer denoting the number of timepoints for the forecast.

x_tt

a design matrix of variables associated to the probability of absence at each site/time.

x_rt

a design matrix of variables associated to recruitment.

x_mt

a design matrix of variables associated to survival.

x_mpast

a design matrix of variables associated to survival from the last year of the "training" perior (i.e., the period using to fit the model and obtain the parameters' estimates).

f_test

a matrix informing the instantaneous fishing mortality rates at each age (columns) and timepoint (rows).

seed

a seed used for the forecasts. Forecasts are obtained through Monte Carlo samples from the posterior predictive distribution. Therefore, a seed is needed to ensure the results' reproducibility.

cores

number of threads used for the forecast. If four chains were used in the drm, then four (or less) threads are recommended.

Value

an object of class "CmdStanGQ" containing samples for the posterior predictive distribution for forecasting.

Details

It is important that the rows of design matrices x_tt, x_rt and x_mt are associated to the same patch/site and timepoints. In addition, the current version of the code assumes the data where forecasts are needed is ordered by "patch" and "site" and, in addition, its patches MUST be the same as the ones used to obtain the parameters' estimates from the the drm object.

Author

lcgodoy