Relationships with covariates

Relationships with covariates

Usage

marg_rec(drm, newdata)

marg_pabs(drm, newdata)

marg_surv(drm, newdata)

Arguments

drm

the output of a fit_drm() call.

newdata

a data.frame with the values of the environmental variables at which we wish to estimate the recruitment, survival, or absence probabilities.

Value

a data.frame with the samples the recruitment (or survival, or absence probabilities) posterior distribution.

Details

the marg_rec function works to obtain relationships regarding recruitment, while marg_surv and marg_abs evaluate relationships with survival and absence probability, respectively.

Author

lcgodoy